BACKPORT: loop: Refactor loop_set_status() size calculation
figure_loop_size() calculates the loop size based on the passed in
parameters, but at the same time it updates the offset and sizelimit
parameters in the loop device configuration. That is a somewhat
unexpected side effect of a function with this name, and it is only only
needed by one of the two callers of this function - loop_set_status().

Move the lo_offset and lo_sizelimit assignment back into loop_set_status(),
and use the newly factored out functions to validate and apply the newly
calculated size. This allows us to get rid of figure_loop_size() in a
follow-up commit.

Signed-off-by: Martijn Coenen <maco@android.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Lisa Liu <liulisa@google.com>
(cherry picked from commit b0bd158dd630bd47640e0e418c062cda1e0da5ad)
Bug: 148607611
Change-Id: Ib8e10eb734c3f95f3c433c13012b378373336d10
Signed-off-by: Euphilia Magenta <xtrffjdev@gmail.com>