blk-mq: Do not wake idle CPUs to queue same-origin requests
Same-origin requests are great for cache locality, but in the case where
the origin CPU is already idle (i.e. the task his migrated), waking it
to process a block request is inefficient for power management. Since
the currently executing CPU is not considered idle, it would be more
power efficient to miss the cache hit and avoid waking an idle CPU.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Euphilia Magenta <xtrffjdev@gmail.com>